pretrain/pretrain_data_process/requirements.txt
addict==2.4.0
aiofiles==23.1.0
anyio==3.6.2
apted==1.0.3
arabic-reshaper==3.0.0
blend-modes==2.1.0
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==3.0.1
ci-info==0.3.0
click==8.1.3
configobj==5.0.8
configparser==5.3.0
contourpy==1.0.7
cycler==0.11.0
datasets==1.6.2
dill==0.3.6
Distance==0.1.3
etelemetry==0.3.0
filelock==3.9.0
fitz==0.0.1.dev2
fonttools==4.38.0
frontend==0.0.3
fsspec==2023.5.0
future==0.18.3
h11==0.14.0
httplib2==0.22.0
huggingface-hub==0.12.1
idna==3.4
imageio==2.26.0
imgaug==0.4.0
importlib-resources==5.12.0
isodate==0.6.1
itsdangerous==2.1.2
kiwisolver==1.4.4
lazy_loader==0.1
looseversion==1.1.2
lxml==4.9.2
matplotlib==3.7.0
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
multiprocess==0.70.14
networkx==3.0
nibabel==5.1.0
nipype==1.8.6
numpy==1.23.5
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72
packaging==23.0
pandas==2.0.1
pathlib==1.0.1
Pillow==9.4.0
Polygon3==3.0.9.1
protobuf==3.20.3
prov==2.0.0
pyarrow==12.0.0
pycocotools==2.0.6
pydot==1.4.2
pygame==2.2.0
PyMuPDF==1.22.2
pyparsing==3.0.9
python-bidi==0.4.2
python-dateutil==2.8.2
pytweening==1.0.4
pytz==2023.3
PyWavelets==1.4.1
pyxnat==1.5
PyYAML==6.0
rdflib==6.3.2
regex==2022.10.31
requests==2.28.2
scikit-image==0.20.0
scipy==1.9.1
seaborn==0.12.2
sentencepiece==0.1.97
shapely==2.0.1
simplejson==3.19.1
sniffio==1.3.0
soupsieve==2.4.1
starlette==0.26.1
synthtiger==1.2.1
terminaltables==3.1.10
tifffile==2023.2.28
tokenizers==0.13.2
torch==1.7.1
torchaudio==0.7.0a0+a853dff
torchvision==0.8.2
tqdm==4.49.0
traits==6.3.2
transformers==4.28.0
tzdata==2023.3
urllib3==1.26.14
uvicorn==0.22.0
xxhash==3.2.0
yapf==0.32.0
zipp==3.15.0
